Cleanup search actions
